♻️ Move printing to utils
