+++ src/Solver.cpp
@@ -7,52 +7,10 @@
 #include <utility>
 #include <vector>
 #include <map>
+#include "Utils.h"
 
 using namespace std;
-
-char go = '-', nogo = 'O';
-
-struct node
-{
-    pair<int, int> loc;
-    pair<int, int> last;
-    vector<pair<int, int>> connections;
-    int cost;
-};
-
-int convertor(char ch)
-{
-    if (ch == go)
-        return 0;
-    else if (ch == nogo)
-        return 1;
-    else
-        throw;
-}
-
-void showmaze2d(vector<vector<int>> maze)
-{
-    for (int i = 0; i < maze.size(); i++)
-    {
-        for (int j = 0; j < maze[i].size(); j++)
-            cout << maze[i][j] << " ";
-        cout << endl;
-    }
-}
-
-void showmazeg(vector<node> mazeg)
-{
-    for (int i = 0; i < mazeg.size(); i++)
-    {
-        node localnode = mazeg[i];
-        cout << localnode.loc.first << "," << localnode.loc.second;
-        for (int k = 0; k < localnode.connections.size(); k++)
-        {
-            cout << " ->" << localnode.connections[k].first << "," << localnode.connections[k].second;
-        }
-        cout << endl;
-    }
-}
+convertor chartoint('-', 'O');
 
 vector<vector<int>> readmazefromfile(char *filename)
 {
@@ -64,7 +22,7 @@ vector<vector<int>> readmazefromfile(char *filename)
         vector<int> buf;
         for (auto &ch : data)
         {
-            buf.push_back(convertor(ch));
+            buf.push_back(chartoint.dothis(ch));
         }
         mazeAs2dArray.push_back(buf);
     }
